Adds defaultTopicTrigger helper
