+++ lib/helpers/index.js
@@ -4,6 +4,7 @@ const cache = require('./cache');
 const campaign = require('./campaign');
 const campaignActivity = require('./campaignActivity');
 const command = require('./command');
+const defaultTopicTrigger = require('./defaultTopicTrigger');
 const request = require('./request');
 const topic = require('./topic');
 const util = require('./util');
@@ -12,6 +13,7 @@ module.exports = {
   cache,
   campaign,
   campaignActivity,
+  defaultTopicTrigger,
   command,
   request,
   topic,
