+++ lib/helpers/topic.js
@@ -23,25 +23,6 @@ function getById(topicId) {
 }
 
 /**
- * @return {Promise}
- */
-function fetchDefaultTopicTriggers() {
-  return contentful.fetchDefaultTopicTriggers()
-    .then(contentfulEntries => contentfulEntries.map(module.exports
-      .parseDefaultTopicTriggerFromContentfulEntry));
-}
-
-/**
- * @param {String} topicId
- * @return {Promise}
- */
-function fetchDefaultTopicTriggersByTopicId(topicId) {
-  return contentful.fetchByResponseReferenceContentfulId(topicId)
-    .then(contentfulEntries => contentfulEntries.map(module.exports
-      .parseDefaultTopicTriggerFromContentfulEntry));
-}
-
-/**
  * @param {String} campaignId
  * @return {Promise}
  */
@@ -136,37 +117,6 @@ function getPostTypeFromContentType(contentType) {
 }
 
 /**
- * Parses a defaultTopicTrigger Contentful entry as data for writing Rivescript to define
- * triggers on the default topic.
- *
- * @param {Object} contentfulEntry - defaultTopicTrigger
- * @return {String}
- */
-function parseDefaultTopicTriggerFromContentfulEntry(contentfulEntry) {
-  const data = {
-    trigger: contentful.getTriggerTextFromDefaultTopicTrigger(contentfulEntry),
-  };
-
-  const responseEntry = contentful.getResponseEntryFromDefaultTopicTrigger(contentfulEntry);
-  if (!responseEntry) {
-    return null;
-  }
-
-  if (contentful.isDefaultTopicTrigger(responseEntry)) {
-    data.redirect = contentful.getTriggerTextFromDefaultTopicTrigger(responseEntry);
-    return data;
-  }
-
-  if (contentful.isMessage(responseEntry)) {
-    data.reply = contentful.getTextFromMessage(responseEntry);
-    return data;
-  }
-
-  data.topicId = contentful.getContentfulIdFromContentfulEntry(responseEntry);
-  return data;
-}
-
-/**
  * @param {Object} contentfulEntry
  * @return {Promise}
  */
@@ -186,9 +136,8 @@ function parseTopicFromContentfulEntry(contentfulEntry) {
   data.templates = Object.assign(topicTemplates, campaignTemplates);
   logger.debug('parseTopicFromContentfulEntry', { data });
 
-  return module.exports.fetchDefaultTopicTriggersByTopicId(topicId)
+  return helpers.defaultTopicTrigger.getByTopicId(topicId)
     .then((triggerObjects) => {
-      // Save linked defaultTopicTriggers as list of the triggers.
       data.defaultTopicTriggers = triggerObjects.map(triggerObject => triggerObject.trigger);
       return helpers.cache.topics.set(topicId, data);
     })
@@ -196,8 +145,6 @@ function parseTopicFromContentfulEntry(contentfulEntry) {
     .catch(err => Promise.reject(err));
 }
 module.exports = {
-  fetchDefaultTopicTriggers,
-  fetchDefaultTopicTriggersByTopicId,
   fetchTopicsByCampaignId,
   getById,
   getDefaultValueFromContentfulEntryAndTemplateName,
@@ -206,6 +153,5 @@ module.exports = {
   getTemplateFromContentfulEntryAndTemplateName,
   getTemplatesFromContentfulEntry,
   getPostTypeFromContentType,
-  parseDefaultTopicTriggerFromContentfulEntry,
   parseTopicFromContentfulEntry,
 };
