+++ lib/middleware/defaultTopicTriggers/index.js
@@ -3,7 +3,7 @@
 const helpers = require('../../helpers');
 
 module.exports = function getDefaultTopicTriggers() {
-  return (req, res) => helpers.topic.fetchDefaultTopicTriggers()
+  return (req, res) => helpers.defaultTopicTrigger.fetchAll()
     .then(data => res.send({ data }))
     .catch(err => helpers.sendErrorResponse(res, err));
 };
